+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 ### Added
 - Added the `Use YouTube nocookie?` plugin setting, which will force usage of the `youtube-nocookie.com` domain for YouTube iframes when enabled
 - Added the `Disable tracking on Vimeo iframes?` plugin setting, which will force usage of the `dnt=1` query parameter for Vimeo iframes when enabled
-- Added `spicyweb\embeddedassets\models\EmbeddedAsset::getIframeCode()`
-- Added `spicyweb\embeddedassets\models\EmbeddedAsset::getIframeSrc()`
+- Added `spicyweb\embeddedassets\models\EmbeddedAsset::getIframeCode()` (like `getVideoCode()`, but for all cases where the embedded asset's `code` is an `<iframe>`)
+- Added `spicyweb\embeddedassets\models\EmbeddedAsset::getIframeSrc()` (like `getVideoUrl()`, but for all cases where the embedded asset's `code` is an `<iframe>`)
 
 ### Changed
 - Embedded Assets now requires Craft 3.6.0 or later